changed type of price to int in game config widget

+++ projet/client_leger/polyquiz/lib/models/game_info_interface.dart
@@ -17,7 +17,7 @@ class GameConfig {
   final String? gameType;
   final bool? private;
   final String? onGoing;
-  final double? price;
+  final int? price;
   final bool? friendsOnly;
   final double? prestige;
